pyproject.toml
[build-system]
requires = ["setuptools>=65.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "media-indexer"
version = "0.1.0"
description = "GPU-accelerated image analysis tool for extracting metadata, faces, objects, and poses"
readme = "README.md"
requires-python = ">=3.10"
license = "MIT"
authors = [
    {name = "Cursor.sh User", email = "user@example.com"}
]
keywords = ["image-analysis", "exif", "face-recognition", "object-detection", "pose-detection"]
classifiers = [
    "Development Status :: 3 - Alpha",
    "Intended Audience :: Developers",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
    "Topic :: Multimedia :: Graphics",
    "Topic :: Scientific/Engineering :: Image Recognition",
]

dependencies = [
    "torch>=2.0.0",
    "torchvision>=0.15.0",
    "numpy>=1.24.0",
    "opencv-python>=4.8.0",
    "Pillow>=10.0.0",
    "rich>=13.0.0",
    "pyyaml>=6.0",
    "face_recognition>=1.3.0",
    "deepface>=0.0.92",
    "insightface>=0.7.3",
    "onnxruntime-gpu>=1.20.0",
    "ultralytics>=8.3.0",
    "image-sidecar-rust @ git+https://github.com/dapperfu/image-sidecar-rust.git",
    "fast-exif-rs-py @ git+https://github.com/dapperfu/fast-exif-rs-py.git",
    "pony>=0.7.17",
    "xxhash>=3.0.0",
    "rawpy>=0.21.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.4.0",
    "pytest-cov>=4.1.0",
    "mypy>=1.5.0",
    "ruff>=0.1.0",
    "strictdoc>=0.0.51",
    "pydocstyle>=6.0.0",
    "uv>=0.1.0",
]

[project.scripts]
media-indexer = "media_indexer.cli:main"
mi = "media_indexer.cli:main"
media-indexer-download-models = "media_indexer.download_models:main"



[tool.mypy]
python_version = "3.10"
warn_return_any = true
warn_unused_configs = true
disallow_untyped_defs = true
ignore_missing_imports = false
line-length = 120

[tool.ruff]
line-length = 120
target-version = "py310"

[tool.ruff.lint]
select = [
    "E",   # pycodestyle errors
    "W",   # pycodestyle warnings
    "F",   # pyflakes
    "I",   # isort
    "B",   # flake8-bugbear
    "C4",  # flake8-comprehensions
    "UP",  # pyupgrade
    "ARG", # flake8-unused-arguments
    "SIM", # flake8-simplify
    "ICN", # flake8-import-conventions
    "PIE", # flake8-pie
    "TID", # flake8-tidy-imports
    "PERF", # perflint
]
ignore = [
    "E501",  # line too long (handled by formatter)
    "B008",  # do not perform function calls in argument defaults
    "ARG001", # unused function argument
]

[tool.ruff.lint.isort]
known-first-party = ["media_indexer"]


